Fixed an issue with whitespace in the table headers

+++ src/main.rs
@@ -1,8 +1,7 @@
 use crate::SqlQuery::{CreateTable, Drop, Insert, Selection};
-use nom::Parser;
 use nom::{
     branch::alt,
-    bytes::complete::{tag, take_while},
+    bytes::complete::{tag, take_until, take_while},
     character::complete::multispace0,
     character::streaming::space0,
     combinator::{map, opt},
@@ -10,6 +9,7 @@ use nom::{
     sequence::delimited,
     IResult,
 };
+use nom::{AsChar, Parser};
 use std::fs::{remove_file, File, OpenOptions};
 use std::io::Write;
 use std::io::{self, BufRead};
@@ -118,24 +118,27 @@ pub fn parse_insert(input: &str) -> IResult<&str, SqlInsert> {
     // Consume the table name surrounded by whitespace
     let (input, table_name) =
         delimited(multispace0, take_while(char::is_alphanumeric), multispace0)(input)?;
+    println!("{:?}", table_name);
     // Consume the columns in between parenthesis
     let (input, columns) = delimited(
         tag("("),
         separated_list0(tag(",").and(space0), take_while(char::is_alphanumeric)),
         tag(")"),
     )(input)?;
+    println!("{:?}", columns);
     // Consume the whitespace if exist
-    let (input, _) = multispace0(input)?;
+    let (input, _) = space0(input)?;
     // Consume the VALUES statement
     let (input, _) = tag("VALUES")(input)?;
     // Consume the whitespace if exist
-    let (input, _) = multispace0(input)?;
+    let (input, _) = space0(input)?;
     // Consume the values in between the parenthesis
     let (input, values) = delimited(
         tag("("),
         separated_list0(tag(",").and(space0), take_while(char::is_alphanumeric)),
         tag(")"),
     )(input)?;
+    println!("{:?}", values);
     // Consume the semi colon ending
     let (input, _) = tag(";")(input)?;
 
@@ -253,11 +256,15 @@ pub fn insert_into(query: SqlInsert<'_>) {
             Ok(s) => str_columns = s,
             Err(e) => eprintln!("{}", e),
         }
+        // Get rid of whitespace between columns and comma
+        let t: String = str_columns.chars().filter(|c| !c.is_whitespace()).collect();
         // Split the String by the commas
-        let parts = str_columns.split(',');
+        let parts = t.split(',');
         // Then turn the result into a vector of columns
         let vec_columns = parts.collect::<Vec<&str>>();
         // Check if the columns provided in the query match the columns in the Table
+        println!("{:?}", vec_columns);
+        println!("{:?}", query.columns);
         if vec_columns == query.columns {
             write_file
                 .write_all("\n".as_bytes())
@@ -283,7 +290,8 @@ pub fn select_table(query: SqlSelect<'_>) {
         .has_headers(false)
         .from_reader(file);
     // Get the headers
-    let headers = rdr.headers().cloned().unwrap();
+    let mut headers = rdr.headers().cloned().unwrap();
+    headers.trim();
     if query.all_columns.is_some() {
         for result in rdr.records() {
             let record = result.expect("Could not read row");
@@ -316,6 +324,7 @@ pub fn select_table(query: SqlSelect<'_>) {
             }
             i += 1;
         }
+        println!("{:?}", col_index);
         for result in rdr.records() {
             let record = result.expect("Could not read row");
             for j in col_index.clone() {
@@ -328,23 +337,29 @@ pub fn select_table(query: SqlSelect<'_>) {
 
 fn main() {
     let mut input = String::new();
-    while input.trim() != "quit" {
+    loop {
         // Reset Input
         input.clear();
         print!("psql> ");
         io::stdout().flush().expect("Cannot flush stdout");
         match io::stdin().read_line(&mut input) {
-            Ok(_) => match parse_query(input.trim()) {
-                Ok(v) => match v.1 {
-                    CreateTable(c) => create_table(c),
-                    Drop(d) => drop_table(d),
-                    Insert(i) => insert_into(i),
-                    Selection(s) => select_table(s),
-                },
-                Err(e) => {
-                    println!("Error: {:?}", e);
+            Ok(_) => {
+                // if the input is quit then break out of the loop
+                if input.trim() == "quit" {
+                    break;
                 }
-            },
+                match parse_query(input.trim()) {
+                    Ok(v) => match v.1 {
+                        CreateTable(c) => create_table(c),
+                        Drop(d) => drop_table(d),
+                        Insert(i) => insert_into(i),
+                        Selection(s) => select_table(s),
+                    },
+                    Err(e) => {
+                        println!("Error: {:?}", e);
+                    }
+                }
+            }
             Err(e) => println!("Error: {:?}", e),
         }
     }
